+++ test/thor_worker.cc
@@ -1,10 +1,10 @@
 #include "test.h"
+#include "utils.h"
 
 #include "midgard/logging.h"
 #include "thor/attributes_controller.h"
 #include "thor/worker.h"
 #include "tyr/actor.h"
-#include <boost/property_tree/ptree.hpp>
 #include <thread>
 #include <unistd.h>
 
@@ -14,16 +14,8 @@ using namespace valhalla::thor;
 
 namespace {
 
-boost::property_tree::ptree json_to_pt(const std::string& json) {
-  std::stringstream ss;
-  ss << json;
-  boost::property_tree::ptree pt;
-  rapidjson::read_json(ss, pt);
-  return pt;
-}
-
 // fake config
-const auto conf = json_to_pt(R"({
+const auto conf = test::json_to_pt(R"({
     "mjolnir":{"tile_dir":"test/data/utrecht_tiles", "concurrency": 1},
     "loki":{
       "actions":["locate","route","sources_to_targets","optimized_route","isochrone","trace_route","trace_attributes"],
@@ -58,7 +50,7 @@ const auto conf = json_to_pt(R"({
 TEST(ThorWorker, test_parse_filter_attributes_defaults) {
   tyr::actor_t actor(conf, true);
 
-  auto result = json_to_pt(actor.trace_attributes(
+  auto result = test::json_to_pt(actor.trace_attributes(
       R"({"costing":"auto","shape_match":"map_snap","shape":[
           {"lat":52.09110,"lon":5.09806},
           {"lat":52.09098,"lon":5.09679}]})"));
@@ -98,7 +90,7 @@ TEST(ThorWorker, test_parse_filter_attributes_excludes) {
   std::vector<std::string> excluded_keys = {"shape", "trip.legs..shape", "trip.legs..shape"};
 
   for (size_t i = 0; i < test_cases.size(); ++i) {
-    auto result = json_to_pt(test_cases[i]);
+    auto result = test::json_to_pt(test_cases[i]);
     EXPECT_FALSE(result.get_child_optional(excluded_keys[i]))
         << "Expected excluded shape | found " + excluded_keys[i] + "=" +
                result.get<std::string>(excluded_keys[i]);
@@ -127,7 +119,7 @@ TEST(ThorWorker, test_parse_filter_attributes_includes) {
   std::vector<std::string> included_keys = {"shape", "trip.legs..shape", "trip.legs..shape"};
 
   for (size_t i = 0; i < test_cases.size(); ++i) {
-    auto result = json_to_pt(test_cases[i]);
+    auto result = test::json_to_pt(test_cases[i]);
     EXPECT_TRUE(result.get_child_optional(included_keys[i]))
         << "Expected " + included_keys[i] + " to be present";
   }
@@ -148,7 +140,7 @@ TEST(ThorWorker, test_linear_references) {
   };
   tyr::actor_t actor(conf, true);
   for (const auto& request : requests) {
-    auto result = json_to_pt(actor.route(request));
+    auto result = test::json_to_pt(actor.route(request));
     const auto& references = result.get_child("trip.linear_references");
     EXPECT_EQ(references.size(), 4);
     for (const auto& reference : references) {