+++ CHANGELOG.md
@@ -142,6 +142,7 @@
    * CHANGED: Speed up parseways stage by avoiding multiple string comparisons [#2518](https://github.com/valhalla/valhalla/pull/2518)
    * CHANGED: Speed up enhance stage by avoiding GraphTileBuilder copying [#2468](https://github.com/valhalla/valhalla/pull/2468)
    * ADDED: Incidents in intersections [#2547](https://github.com/valhalla/valhalla/pull/2547)
+   * CHANGED: Refactor mapmatching configuration to use a struct (instead of `boost::property_tree::ptree`). [#2485](https://github.com/valhalla/valhalla/pull/2485)
 
 ## Release Date: 2019-11-21 Valhalla 3.0.9
 * **Bug Fix**